requirements.txt
beautifulsoup4==4.8.0
certifi==2019.6.16
chardet==3.0.4
cssselect==1.0.3
Django==2.2.3
djangorestframework==3.10.2
feedfinder2==0.0.4
feedparser==5.2.1
idna==2.8
jieba3k==0.35.1
lxml==4.3.4
multidict==4.5.2
newspaper3k==0.2.8
nltk==3.4.4
Pillow==6.1.0
psycopg2-binary==2.8.3
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1.1
requests==2.22.0
requests-file==1.4.3
six==1.12.0
soupsieve==1.9.2
sqlparse==0.3.0
tinysegmenter==0.3
tldextract==2.2.1
urllib3==1.25.3
vcrpy==2.0.1
wrapt==1.11.2
yarl==1.3.0
